requirements.txt
ccxt==3.0.5
python-dotenv==1.0.0
pandas==2.1.3
numpy==1.29.4
aiohttp==3.9.4
pytest==7.4.0
black==24.7.0
ruff==0.0.264
mypy==1.4.1
